requirements.txt
# Core
paho-mqtt==1.6.1
influxdb-client==1.36.0
python-dotenv==1.0.0
schedule==1.2.0

# Data Science (Heavy stuff)
pandas>=1.5.0
numpy>=1.24.0
prophet>=1.1.2
